+++ include/pixels.h
@@ -12,7 +12,7 @@
 
 void setup_pixels(struct app_config *aconf);
 void handle_new_pixel(struct app_config *aconf, uint16_t x, uint16_t y, uint8_t r, uint8_t g, uint8_t b);
-uint32_t *swap_buffers(struct app_config *aconf);
+uint64_t *swap_buffers(struct app_config *aconf);
 std::optional<std::vector<uint8_t>> buffer_to_png(struct app_config *aconf, uint32_t* buffer);
 
 #endif //SIXELPING_RECEIVER_DPDK_PIXELS_H
